+++ packages/notification-services-controller/src/NotificationServicesController/NotificationServicesController.ts
@@ -1,10 +1,9 @@
 import type {
-  RestrictedMessenger,
   ControllerGetStateAction,
   ControllerStateChangeEvent,
   StateMetadata,
-} from '@metamask/base-controller';
-import { BaseController } from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
+import { BaseController } from '@metamask/base-controller/next';
 import {
   isValidHexAddress,
   toChecksumHexAddress,
@@ -17,6 +16,7 @@ import {
   KeyringTypes,
   type KeyringControllerState,
 } from '@metamask/keyring-controller';
+import type { Messenger } from '@metamask/messenger';
 import type { AuthenticationController } from '@metamask/profile-sync-controller';
 import { assert } from '@metamask/utils';
 import log from 'loglevel';
@@ -51,7 +51,7 @@ const controllerName = 'NotificationServicesController';
  */
 export type NotificationServicesControllerState = {
   /**
-   * We store and manage accounts that have been seen/visted through the
+   * We store and manage accounts that have been seen/visited through the
    * account subscription. This allows us to track and add notifications for new accounts and not previous accounts added.
    */
   subscriptionAccountsSeen: string[];
@@ -195,7 +195,7 @@ export type Actions =
   | NotificationServicesControllerDeleteNotificationsById;
 
 // Allowed Actions
-export type AllowedActions =
+type AllowedActions =
   // Keyring Controller Requests
   | KeyringControllerGetStateAction
   // Auth Controller Requests
@@ -231,7 +231,7 @@ export type Events =
   | MarkNotificationsAsReadEvent;
 
 // Allowed Events
-export type AllowedEvents =
+type AllowedEvents =
   // Keyring Events
   | KeyringControllerStateChangeEvent
   | KeyringControllerLockEvent
@@ -241,12 +241,10 @@ export type AllowedEvents =
   | NotificationServicesPushControllerStateChangeEvent;
 
 // Type for the messenger of NotificationServicesController
-export type NotificationServicesControllerMessenger = RestrictedMessenger<
+export type NotificationServicesControllerMessenger = Messenger<
   typeof controllerName,
   Actions | AllowedActions,
-  Events | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  Events | AllowedEvents
 >;
 
 type FeatureAnnouncementEnv = {
@@ -267,12 +265,10 @@ export default class NotificationServicesController extends BaseController<
     isUnlocked: false,
 
     setupLockedStateSubscriptions: (onUnlock: () => Promise<void>) => {
-      const { isUnlocked } = this.messagingSystem.call(
-        'KeyringController:getState',
-      );
+      const { isUnlocked } = this.messenger.call('KeyringController:getState');
       this.#keyringController.isUnlocked = isUnlocked;
 
-      this.messagingSystem.subscribe('KeyringController:unlock', () => {
+      this.messenger.subscribe('KeyringController:unlock', () => {
         this.#keyringController.isUnlocked = true;
         // messaging system cannot await promises
         // we don't need to wait for a result on this.
@@ -280,7 +276,7 @@ export default class NotificationServicesController extends BaseController<
         onUnlock();
       });
 
-      this.messagingSystem.subscribe('KeyringController:lock', () => {
+      this.messenger.subscribe('KeyringController:lock', () => {
         this.#keyringController.isUnlocked = false;
       });
     },
@@ -288,15 +284,15 @@ export default class NotificationServicesController extends BaseController<
 
   readonly #auth = {
     getBearerToken: async () => {
-      return await this.messagingSystem.call(
+      return await this.messenger.call(
         'AuthenticationController:getBearerToken',
       );
     },
     isSignedIn: () => {
-      return this.messagingSystem.call('AuthenticationController:isSignedIn');
+      return this.messenger.call('AuthenticationController:isSignedIn');
     },
     signIn: async () => {
-      return await this.messagingSystem.call(
+      return await this.messenger.call(
         'AuthenticationController:performSignIn',
       );
     },
@@ -309,13 +305,13 @@ export default class NotificationServicesController extends BaseController<
     isSetup: false,
 
     subscribeToPushNotifications: async () => {
-      await this.messagingSystem.call(
+      await this.messenger.call(
         'NotificationServicesPushController:subscribeToPushNotifications',
       );
     },
     enablePushNotifications: async (addresses: string[]) => {
       try {
-        await this.messagingSystem.call(
+        await this.messenger.call(
           'NotificationServicesPushController:enablePushNotifications',
           addresses,
         );
@@ -325,7 +321,7 @@ export default class NotificationServicesController extends BaseController<
     },
     disablePushNotifications: async () => {
       try {
-        await this.messagingSystem.call(
+        await this.messenger.call(
           'NotificationServicesPushController:disablePushNotifications',
         );
       } catch (e) {
@@ -333,7 +329,7 @@ export default class NotificationServicesController extends BaseController<
       }
     },
     subscribe: () => {
-      this.messagingSystem.subscribe(
+      this.messenger.subscribe(
         'NotificationServicesPushController:onNewNotifications',
         (notification) => {
           // eslint-disable-next-line @typescript-eslint/no-floating-promises
@@ -372,9 +368,7 @@ export default class NotificationServicesController extends BaseController<
     isNotificationAccountsSetup: false,
 
     getNotificationAccounts: () => {
-      const { keyrings } = this.messagingSystem.call(
-        'KeyringController:getState',
-      );
+      const { keyrings } = this.messenger.call('KeyringController:getState');
       const firstHDKeyring = keyrings.find(
         (k) => k.type === KeyringTypes.hd.toString(),
       );
@@ -451,7 +445,7 @@ export default class NotificationServicesController extends BaseController<
      * And call effects to subscribe/unsubscribe to notifications.
      */
     subscribe: () => {
-      this.messagingSystem.subscribe(
+      this.messenger.subscribe(
         'KeyringController:stateChange',
         // eslint-disable-next-line @typescript-eslint/no-misused-promises
         async (totalAccounts, prevTotalAccounts) => {
@@ -535,22 +529,22 @@ export default class NotificationServicesController extends BaseController<
   }
 
   #registerMessageHandlers(): void {
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateMetamaskNotificationsList`,
       this.updateMetamaskNotificationsList.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:disableNotificationServices`,
       this.disableNotificationServices.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:getNotificationsByType`,
       this.getNotificationsByType.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:deleteNotificationsById`,
       this.deleteNotificationsById.bind(this),
     );
@@ -1012,7 +1006,7 @@ export default class NotificationServicesController extends BaseController<
         state.metamaskNotificationsList = metamaskNotifications;
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:notificationsListUpdated`,
         this.state.metamaskNotificationsList,
       );
@@ -1093,7 +1087,7 @@ export default class NotificationServicesController extends BaseController<
       await this.deleteNotificationById(id);
     }
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:notificationsListUpdated`,
       this.state.metamaskNotificationsList,
     );
@@ -1208,7 +1202,7 @@ export default class NotificationServicesController extends BaseController<
       );
     });
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:markNotificationsAsRead`,
       this.state.metamaskNotificationsList,
     );
@@ -1246,7 +1240,7 @@ export default class NotificationServicesController extends BaseController<
         }
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:notificationsListUpdated`,
         this.state.metamaskNotificationsList,
       );